Replace test role data
