--- backend/src/utils/role/default.js
@@ -1,32 +1,0 @@
-export const adminRole = {
-    profile: {
-        all: ['read', 'update', 'delete'],
-        params: {
-            password: ['!read', 'update'],
-        },
-    },
-    board: {
-        all: ['read', 'write'],
-    },
-}
-
-export const userRole = {
-    profile: {
-        all: {
-            any: ['read'],
-            own: ['read', 'update', 'delete'],
-        },
-        params: {
-            password: {
-                any: ['!read'],
-                own: ['!read'],
-            },
-        },
-    },
-    board: {
-        all: ['read', 'write'],
-        params: {
-            '12': ['!write'],
-        },
-    },
-}
